+++ yandex/resource_yandex_lockbox_secret_iam_binding.go
@@ -1,0 +1,7 @@
+package yandex
+
+import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+func resourceYandexLockboxSecretIAMBinding() *schema.Resource {
+	return resourceIamBindingWithImport(IamLockboxSecretSchema, newLockboxSecretIamUpdater, LockboxSecretIDParseFunc)
+}