+++ CHANGELOG.md
@@ -7,6 +7,7 @@ FEATURES:
 * **New Data Source:** `yandex_kms_symmetric_key`
 * storage: added bucket and object `tagging`
 * k8s: add `container_network` attribute with `pod_mtu` to `node_group` resource and data source
+* **New Resource** `yandex_lockbox_secret_iam_binding`
 
 BUG FIXES:
 * organizationmanager: data source `yandex_organizationmanager_saml_federation_user_account` account creation behaviour deprecation notice added